done changes in URI of controllers
